requirements.txt
fastapi==0.95.2
flask
requests
python-dotenv
poe-api-wrapper
supabase
uvicorn
gunicorn
